+++ core/src/components/menu-controller/menu-controller.ts
@@ -1,25 +1,19 @@
 import { Build, Component, Method } from '@stencil/core';
 
-import { config } from '../../global/config';
-import { AnimationBuilder, IonicAnimation, MenuControllerI, MenuI } from '../../interface';
-
-import { menuOverlayAnimation } from './animations/overlay';
-import { menuPushAnimation } from './animations/push';
-import { menuRevealAnimation } from './animations/reveal';
+import { AnimationBuilder, IonicAnimation, MenuI } from '../../interface';
+import { menuController } from '../../utils/menu-controller';
 
 @Component({
   tag: 'ion-menu-controller',
   styleUrl: 'menu-controller.scss'
 })
-export class MenuController implements MenuControllerI {
-
-  private menus: MenuI[] = [];
-  private menuAnimations = new Map<string, ((menu: MenuI) => IonicAnimation) | AnimationBuilder>();
+export class MenuController {
 
   constructor() {
-    this.registerAnimation('reveal', menuRevealAnimation);
-    this.registerAnimation('push', menuPushAnimation);
-    this.registerAnimation('overlay', menuOverlayAnimation);
+    if (Build.isDev) {
+      console.warn(`[DEPRECATED][ion-menu-controller] Use the menuController export from @ionic/core:
+  import { menuController } from '@ionic/core';`);
+    }
   }
 
   /**
@@ -32,12 +26,8 @@ export class MenuController implements MenuControllerI {
    * @param menu The menuId or side of the menu to open.
    */
   @Method()
-  async open(menu?: string | null): Promise<boolean> {
-    const menuEl = await this.get(menu);
-    if (menuEl) {
-      return menuEl.open();
-    }
-    return false;
+  open(menu?: string | null) {
+    return menuController.open(menu);
   }
 
   /**
@@ -48,12 +38,8 @@ export class MenuController implements MenuControllerI {
    * @param menu The menuId or side of the menu to close.
    */
   @Method()
-  async close(menu?: string | null): Promise<boolean> {
-    const menuEl = await (menu !== undefined ? this.get(menu) : this.getOpen());
-    if (menuEl !== undefined) {
-      return menuEl.close();
-    }
-    return false;
+  close(menu?: string | null) {
+    return menuController.close(menu);
   }
 
   /**
@@ -64,12 +50,8 @@ export class MenuController implements MenuControllerI {
    * @param menu The menuId or side of the menu to toggle.
    */
   @Method()
-  async toggle(menu?: string | null): Promise<boolean> {
-    const menuEl = await this.get(menu);
-    if (menuEl) {
-      return menuEl.toggle();
-    }
-    return false;
+  toggle(menu?: string | null) {
+    return menuController.toggle(menu);
   }
 
   /**
@@ -83,12 +65,8 @@ export class MenuController implements MenuControllerI {
    * @param menu The menuId or side of the menu to enable or disable.
    */
   @Method()
-  async enable(enable: boolean, menu?: string | null): Promise<HTMLIonMenuElement | undefined> {
-    const menuEl = await this.get(menu);
-    if (menuEl) {
-      menuEl.disabled = !enable;
-    }
-    return menuEl;
+  enable(enable: boolean, menu?: string | null) {
+    return menuController.enable(enable, menu);
   }
 
   /**
@@ -98,12 +76,8 @@ export class MenuController implements MenuControllerI {
    * @param menu The menuId or side of the menu to enable or disable the swipe gesture on.
    */
   @Method()
-  async swipeGesture(enable: boolean, menu?: string | null): Promise<HTMLIonMenuElement | undefined> {
-    const menuEl = await this.get(menu);
-    if (menuEl) {
-      menuEl.swipeGesture = enable;
-    }
-    return menuEl;
+  swipeGesture(enable: boolean, menu?: string | null) {
+    return menuController.swipeGesture(enable, menu);
   }
 
   /**
@@ -114,14 +88,8 @@ export class MenuController implements MenuControllerI {
    * @param menu The menuId or side of the menu that is being checked.
    */
   @Method()
-  async isOpen(menu?: string | null): Promise<boolean> {
-    if (menu != null) {
-      const menuEl = await this.get(menu);
-      return (menuEl !== undefined && menuEl.isOpen());
-    } else {
-      const menuEl = await this.getOpen();
-      return menuEl !== undefined;
-    }
+  isOpen(menu?: string | null) {
+    return menuController.isOpen(menu);
   }
 
   /**
@@ -132,12 +100,8 @@ export class MenuController implements MenuControllerI {
    * @param menu The menuId or side of the menu that is being checked.
    */
   @Method()
-  async isEnabled(menu?: string | null): Promise<boolean> {
-    const menuEl = await this.get(menu);
-    if (menuEl) {
-      return !menuEl.disabled;
-    }
-    return false;
+  isEnabled(menu?: string | null) {
+    return menuController.isEnabled(menu);
   }
 
   /**
@@ -149,63 +113,24 @@ export class MenuController implements MenuControllerI {
    * @param menu The menuId or side of the menu.
    */
   @Method()
-  async get(menu?: string | null): Promise<HTMLIonMenuElement | undefined> {
-    if (Build.isDev) {
-      if (menu === 'left') {
-        console.error('menu.side=left is deprecated, use "start" instead');
-        return undefined;
-      }
-      if (menu === 'right') {
-        console.error('menu.side=right is deprecated, use "end" instead');
-        return undefined;
-      }
-    }
-    await this.waitUntilReady();
-
-    if (menu === 'start' || menu === 'end') {
-      // there could be more than one menu on the same side
-      // so first try to get the enabled one
-      const menuRef = this.find(m => m.side === menu && !m.disabled);
-      if (menuRef) {
-        return menuRef;
-      }
-
-      // didn't find a menu side that is enabled
-      // so try to get the first menu side found
-      return this.find(m => m.side === menu);
-
-    } else if (menu != null) {
-      // the menuId was not left or right
-      // so try to get the menu by its "id"
-      return this.find(m => m.menuId === menu);
-    }
-
-    // return the first enabled menu
-    const menuEl = this.find(m => !m.disabled);
-    if (menuEl) {
-      return menuEl;
-    }
-
-    // get the first menu in the array, if one exists
-    return this.menus.length > 0 ? this.menus[0].el : undefined;
+  get(menu?: string | null) {
+    return menuController.get(menu);
   }
 
   /**
    * Get the instance of the opened menu. Returns `null` if a menu is not found.
    */
   @Method()
-  async getOpen(): Promise<HTMLIonMenuElement | undefined> {
-    await this.waitUntilReady();
-    return this.getOpenSync();
+  getOpen() {
+    return menuController.getOpen();
   }
 
   /**
    * Get all menu instances.
    */
   @Method()
-  async getMenus(): Promise<HTMLIonMenuElement[]> {
-    await this.waitUntilReady();
-    return this.getMenusSync();
+  getMenus() {
+    return menuController.getMenus();
   }
 
   /**
@@ -213,9 +138,8 @@ export class MenuController implements MenuControllerI {
    * menu is currently animating.
    */
   @Method()
-  async isAnimating(): Promise<boolean> {
-    await this.waitUntilReady();
-    return this.isAnimatingSync();
+  isAnimating() {
+    return menuController.isAnimating();
   }
 
   /**
@@ -227,95 +151,6 @@ export class MenuController implements MenuControllerI {
    */
   @Method()
   async registerAnimation(name: string, animation: AnimationBuilder | ((menu: MenuI) => IonicAnimation)) {
-    this.menuAnimations.set(name, animation);
-  }
-
-  /**
-   * @internal
-   */
-  @Method()
-  _getInstance(): Promise<MenuControllerI> {
-    return Promise.resolve(this);
-  }
-
-  _register(menu: MenuI) {
-    const menus = this.menus;
-    if (menus.indexOf(menu) < 0) {
-      if (!menu.disabled) {
-        this._setActiveMenu(menu);
-      }
-      menus.push(menu);
-    }
-  }
-
-  _unregister(menu: MenuI) {
-    const index = this.menus.indexOf(menu);
-    if (index > -1) {
-      this.menus.splice(index, 1);
-    }
-  }
-
-  _setActiveMenu(menu: MenuI) {
-    // if this menu should be enabled
-    // then find all the other menus on this same side
-    // and automatically disable other same side menus
-    const side = menu.side;
-    this.menus
-      .filter(m => m.side === side && m !== menu)
-      .forEach(m => m.disabled = true);
-  }
-
-  async _setOpen(menu: MenuI, shouldOpen: boolean, animated: boolean): Promise<boolean> {
-    if (this.isAnimatingSync()) {
-      return false;
-    }
-    if (shouldOpen) {
-      const openedMenu = await this.getOpen();
-      if (openedMenu && menu.el !== openedMenu) {
-        await openedMenu.setOpen(false, false);
-      }
-    }
-    return menu._setOpen(shouldOpen, animated);
-  }
-
-  async _createAnimation(type: string, menuCmp: MenuI): Promise<any> {
-    const animationBuilder = this.menuAnimations.get(type) as any;
-    if (!animationBuilder) {
-      throw new Error('animation not registered');
-    }
-
-    const animation = animationBuilder(menuCmp);
-
-    if (!config.getBoolean('animated', true)) {
-      animation.duration(0);
-    }
-    return animation;
-  }
-
-  getOpenSync(): HTMLIonMenuElement | undefined {
-    return this.find(m => m._isOpen);
-  }
-
-  getMenusSync(): HTMLIonMenuElement[] {
-    return this.menus.map(menu => menu.el);
-  }
-
-  isAnimatingSync(): boolean {
-    return this.menus.some(menu => menu.isAnimating);
-  }
-
-  private find(predicate: (menu: MenuI) => boolean): HTMLIonMenuElement | undefined {
-    const instance = this.menus.find(predicate);
-    if (instance !== undefined) {
-      return instance.el;
-    }
-    return undefined;
-  }
-
-  private waitUntilReady() {
-    return Promise.all(
-      Array.from(document.querySelectorAll('ion-menu'))
-        .map(menu => menu.componentOnReady())
-    );
+    return menuController.registerAnimation(name, animation);
   }
 }