+++ vue/src/app-initialize.ts
@@ -4,12 +4,12 @@ import { IonicConfig } from '@ionic/core';
 
 // Webpack import for ionicons
 import { addIcons } from 'ionicons';
-import { close, reorder, menu, arrowDown, arrowForward, arrowBack, search, closeCircle } from 'ionicons/icons';
+import { arrowBack, arrowDown, arrowForward, close, closeCircle, menu, reorder, search } from 'ionicons/icons';
 
 // import '@ionic/core/css/ionic.bundle.css';
 // import 'ionicons/dist/collection/icon/icon.css';
 
-import { defineCustomElements } from '@ionic/core/loader';
+import { applyPolyfills, defineCustomElements } from '@ionic/core/loader';
 import { IonicWindow } from './interfaces';
 
 export function appInitialize(config?: IonicConfig) {
@@ -17,7 +17,8 @@ export function appInitialize(config?: IonicConfig) {
   const Ionic = (win.Ionic = win.Ionic || {});
 
   Ionic.config = config;
-  defineCustomElements(window);
+  applyPolyfills().then(() => defineCustomElements(win));
+
   // Icons that are used by internal components
   addIcons({
     'ios-close': close.ios,